Finished MyFlights Screen
Add list of reservations, button to add new reservation and go to
booking.
Get the user reservation from firestore and render in the list items.
