+++ VolcanAirApp/src/components/MyFligths/FlightItem.tsx
@@ -3,12 +3,13 @@ import React from 'react'
 import myFlightsQueries from '../../Queries/MyFligths/myFlightsQueries'
 import { styleMyFligths } from './styleMyFligths';
 import FontAwesome5 from 'react-native-vector-icons/FontAwesome5';
+
 type flightProps={
     flight: Flight;
 }
 
 type Flight = {
-    date: string;
+    date: any;
     destination_city: City;
     origin_city: City;
     no_passengers: number;
@@ -19,11 +20,20 @@ type City={
     name: string;
 }
 
+const formatDate = (timestamp: any) => {
+  const date = timestamp.toDate();
+  const day = date.getDate();
+  const month = date.toLocaleString('default', { month: 'long' }); // Obtener el nombre del mes
+  const year = date.getFullYear();
+  // Formatear la fecha como una cadena escrita
+  return `${day} de ${month} de ${year}`;
+};
+
 export default function FlightItem({flight}: flightProps): JSX.Element {
   const icon = <FontAwesome5 name={'plane'} color={"#5c6df8"} size={20} style={{textAlign:'center'}}/>; 
+   
   return (
     <View  style={styleMyFligths.ContainerFLatList}>
-      <View>
         <View style={styleMyFligths.names}>
           <Text style={styleMyFligths.originCity}>{flight.origin_city.code}</Text>
           <View style={styleMyFligths.icon}>{icon}</View>
@@ -37,11 +47,10 @@ export default function FlightItem({flight}: flightProps): JSX.Element {
         </View>
         <View style={styleMyFligths.lineaBlack}>
           <View style={styleMyFligths.containertLineaBLack}>
-            <Text style={styleMyFligths.date}>{flight.date}</Text>
+            <Text style={styleMyFligths.date}>{formatDate(flight.date)}</Text>
             <Text style={styleMyFligths.passengers}>{flight.no_passengers} passengers</Text>
           </View>
         </View>
-      </View>
   </View>
   )
 }
