+++ VolcanAirApp/src/styles/Components/MyFlights/FlightList.tsx
@@ -1,0 +1,9 @@
+import { StyleSheet } from 'react-native'
+import React from 'react'
+
+export const StyleFlightList = StyleSheet.create({
+    flistFlights: {
+        height: 680,
+    }
+})
+
