+++ VolcanAirApp/src/styles/Components/SignUp/StyleCheckBoxSubs.tsx
@@ -1,0 +1,13 @@
+import { StyleSheet } from 'react-native'
+import React from 'react'
+
+export const StyleCheckBoxSubscribe = StyleSheet.create({
+    checkBoxContainer: {
+        marginTop: 4,
+        display: 'flex',
+        flexDirection: 'row',
+        alignItems: 'center',
+        marginLeft: 5
+    },
+})
+
