+++ VolcanAirApp/src/styles/Components/SignUp/StyleCheckBoxTerms.tsx
@@ -1,0 +1,19 @@
+import { StyleSheet } from 'react-native'
+import React from 'react'
+
+export const StyleCheckBoxTerms = StyleSheet.create({
+    checkBoxContainer: {
+        marginTop: 4,
+        display: 'flex',
+        flexDirection: 'row',
+        alignItems: 'center',
+        marginLeft: 5
+    },
+    underline: {
+        textDecorationLine: 'underline',
+    },
+    asteristik: {
+        color: 'red',
+    }
+})
+
