+++ VolcanAirApp/src/styles/Components/SignUp/StyleInputSignUp.tsx
@@ -1,0 +1,12 @@
+import { StyleSheet } from 'react-native'
+import React from 'react'
+
+export const StyleInputSignUp = StyleSheet.create({
+    input: {
+        marginTop: 2,
+        borderWidth: 0.8,
+        borderColor: '#474747',
+        width: '110%',
+    },
+})
+
